+++ paddle/phi/api/yaml/legacy_api.yaml
@@ -1501,6 +1501,14 @@
     func : matmul
   backward : matmul_grad
 
+- api : matrix_nms
+  args : (Tensor bboxes, Tensor scores, float score_threshold, int nms_top_k, int keep_top_k, float post_threshold=0., bool use_gaussian = false, float gaussian_sigma = 2.0, int background_label = 0, bool normalized = true)
+  output : Tensor(out), Tensor(index), Tensor(roisnum)
+  infer_meta :
+    func : MatrixNMSInferMeta
+  kernel :
+    func : matrix_nms
+
 # matrix_power
 - api : matrix_power
   args : (Tensor x, int n)
