+++ paddle/phi/kernels/matrix_nms_kernel.h
@@ -1,0 +1,37 @@
+// Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "paddle/phi/core/dense_tensor.h"
+
+namespace phi {
+
+template <typename T, typename Context>
+void MatrixNMSKernel(const Context& ctx,
+                     const DenseTensor& bboxes,
+                     const DenseTensor& scores,
+                     float score_threshold,
+                     int nms_top_k,
+                     int keep_top_k,
+                     float post_threshold,
+                     bool use_gaussian,
+                     float gaussian_sigma,
+                     int background_label,
+                     bool normalized,
+                     DenseTensor* out,
+                     DenseTensor* index,
+                     DenseTensor* roisnum);
+
+}  // namespace phi
