+++ python/paddle/fluid/layers/detection.py
@@ -3642,6 +3642,16 @@ def matrix_nms(bboxes,
                                           keep_top_k=200,
                                           normalized=False)
     """
+    if in_dygraph_mode():
+        attrs = (score_threshold, nms_top_k, keep_top_k, post_threshold,
+                 use_gaussian, gaussian_sigma, background_label, normalized)
+
+        out, index = _C_ops.final_state_matrix_nms(bboxes, scores, *attrs)
+        if return_index:
+            return out, index
+        else:
+            return out
+
     check_variable_and_dtype(bboxes, 'BBoxes', ['float32', 'float64'],
                              'matrix_nms')
     check_variable_and_dtype(scores, 'Scores', ['float32', 'float64'],
@@ -3664,13 +3674,13 @@ def matrix_nms(bboxes,
                          'Scores': scores
                      },
                      attrs={
-                         'background_label': background_label,
                          'score_threshold': score_threshold,
                          'post_threshold': post_threshold,
                          'nms_top_k': nms_top_k,
-                         'gaussian_sigma': gaussian_sigma,
-                         'use_gaussian': use_gaussian,
                          'keep_top_k': keep_top_k,
+                         'use_gaussian': use_gaussian,
+                         'gaussian_sigma': gaussian_sigma,
+                         'background_label': background_label,
                          'normalized': normalized
                      },
                      outputs={
